Create elk.sh
By understanding this ELK script, we can automate our calculation and data analysis. This script can be used for convergence tests for K points, lattice constant optimization, finding the no. of scf loops etc. 
To run this script, a template file('btemp') is needed, With 'msk' variable for the value that we want to change in the automation.